address some comments
