+++ tests/framework/test_time.py
@@ -5,6 +5,8 @@ import pandas as pd
 import pytest
 
 from vivarium.framework.engine import SimulationContext as SimulationContext_
+from vivarium.framework.time import step_size_post_processor
+from vivarium.framework.values import ValuesManager, list_combiner
 
 from .components.mocks import (
     Listener,
@@ -12,6 +14,17 @@ from .components.mocks import (
     MockComponentB,
     MockGenericComponent,
 )
+
+
+@pytest.fixture
+def manager(mocker):
+    manager = ValuesManager()
+    builder = mocker.MagicMock()
+    builder.time.simulant_step_sizes = lambda: lambda: lambda idx: pd.Series(
+        [pd.Timedelta(days=3) if i % 2 == 0 else pd.Timedelta(days=5) for i in idx], index=idx
+    )
+    manager.setup(builder)
+    return manager
 
 
 @pytest.fixture()
@@ -27,6 +40,20 @@ def components():
         Listener("listener"),
         MockComponentB("spoon", "antelope", 23),
     ]
+
+
+def active_simulants(sim):
+    return sim._clock.get_active_population(sim.get_population().index, sim._clock.event_time)
+
+
+def step_pipeline(sim):
+    return sim._values.get_value("simulant_step_size")(
+        sim._population.get_population(True).index
+    )
+
+
+def step_column(sim):
+    return sim._population._population.step_size
 
 
 class StepModifier(MockGenericComponent):
@@ -47,26 +74,22 @@ def test_align_times(SimulationContext, base_config, components):
     sim.setup()
     sim.initialize_simulants()
     pop_size = len(sim.get_population())
-    active_simulants = lambda: sim._clock.get_active_population(
-        sim.get_population().index,
-        sim._clock.event_time,
-    )
     # After initialization, all simulants should be aligned to event times
-    assert len(active_simulants()) == pop_size
+    assert len(active_simulants(sim)) == pop_size
 
     sim.step()
     # After one step (and no step adjustments, simulants should still be aligned)
-    assert len(active_simulants()) == pop_size
+    assert len(active_simulants(sim)) == pop_size
 
     sim._population._population.step_size *= 2
 
     sim.step()
     # No simulants should be aligned after a step size adjustment
-    assert active_simulants().empty
+    assert active_simulants(sim).empty
 
     sim.step()
     # Now they should be aligned again
-    assert len(active_simulants()) == pop_size
+    assert len(active_simulants(sim)) == pop_size
 
 
 def test_unequal_steps(SimulationContext, base_config, components):
@@ -75,15 +98,11 @@ def test_unequal_steps(SimulationContext, base_config, components):
     sim.setup()
     sim.initialize_simulants()
     pop_size = len(sim.get_population())
-    active_simulants = lambda: sim._clock.get_active_population(
-        sim.get_population().index,
-        sim._clock.event_time,
-    )
 
     # Check that the 0th simulant won't step forward
     sim._population._population.step_size[0] *= 2
     sim.step()
-    assert len(active_simulants()) == pop_size - 1
+    assert len(active_simulants(sim)) == pop_size - 1
 
     # Show that now that the next_event_time is updated, 0th simulant isn't included in events
     sim.step()
@@ -91,7 +110,7 @@ def test_unequal_steps(SimulationContext, base_config, components):
         assert 0 not in index
 
     # Check that everybody will step forward next step
-    assert len(active_simulants()) == pop_size
+    assert len(active_simulants(sim)) == pop_size
 
     # Check that they are actually included in events
     sim.step()
@@ -104,7 +123,7 @@ def test_unequal_steps(SimulationContext, base_config, components):
     # Do a step just to update the next_event_time
     sim.step()
     # Check that next step, we will still update all
-    assert len(active_simulants()) == pop_size
+    assert len(active_simulants(sim)) == pop_size
     # Check that we actually include simulant of index 7 in events
     sim.step()
     for index in listener.event_indexes.values():
@@ -115,24 +134,11 @@ def test_step_pipeline(SimulationContext, base_config, components):
     sim = SimulationContext(base_config, components)
     sim.setup()
     sim.initialize_simulants()
-    pop_size = len(sim.get_population())
-    active_simulants = lambda: sim._clock.get_active_population(
-        sim.get_population().index,
-        sim._clock.event_time,
-    )
 
-    pipeline = lambda: sim._values.get_value("simulant_step_size")(
-        sim._population.get_population(True).index
-    )
-
-    column = lambda: sim._population._population.step_size
-
-    assert np.all(pipeline() == column())
-    assert len(active_simulants()) == pop_size
+    assert np.all(step_pipeline(sim) == step_column(sim))
 
     sim.step()
-    assert np.all(pipeline() == column())
-    assert len(active_simulants()) == pop_size
+    assert np.all(step_pipeline(sim) == step_column(sim))
 
 
 @pytest.mark.parametrize("step_modifier", [0.5, 1, 2, 3.5, 5])
@@ -141,18 +147,10 @@ def test_step_pipeline_with_modifier(SimulationContext, base_config, step_modifi
     sim.setup()
     sim.initialize_simulants()
     pop_size = len(sim.get_population())
-    active_simulants = lambda: sim._clock.get_active_population(
-        sim.get_population().index,
-        sim._clock.event_time,
-    )
-    pipeline = lambda: sim._values.get_value("simulant_step_size")(
-        sim._population.get_population(True).index
-    )
-    column = lambda: sim._population._population.step_size
 
     ## Everyone starts active
-    assert np.all(pipeline() == column())
-    assert len(active_simulants()) == pop_size
+    assert np.all(step_pipeline(sim) == step_column(sim))
+    assert len(active_simulants(sim)) == pop_size
 
     ## Go through a couple simulant step cycles
     for _ in range(2):
@@ -160,10 +158,42 @@ def test_step_pipeline_with_modifier(SimulationContext, base_config, step_modifi
             ## Nobody Should update here.
             ## We subtract  a step for the last step of the modified range
             sim.step()
-            assert np.all(pipeline() == column())
-            assert active_simulants().empty
+            assert np.all(step_pipeline(sim) == step_column(sim))
+            assert active_simulants(sim).empty
 
         ## Everyone should update again
         sim.step()
-        assert np.all(pipeline() == column())
-        assert len(active_simulants()) == pop_size
+        assert np.all(step_pipeline(sim) == step_column(sim))
+        assert len(active_simulants(sim)) == pop_size
+
+
+def test_step_size_post_processor(manager):
+    index = pd.Index(range(10))
+
+    pipeline = manager.register_value_producer(
+        "test_step_size",
+        source=lambda idx: [pd.Series(pd.Timedelta(days=2), index=idx)],
+        preferred_combiner=list_combiner,
+        preferred_post_processor=step_size_post_processor,
+    )
+    manager.register_value_modifier(
+        "test_step_size",
+        modifier=lambda idx: pd.Series(
+            [pd.Timedelta(days=7) if i % 2 == 0 else pd.Timedelta(days=5) for i in idx],
+            index=idx,
+        ),
+    )
+    manager.register_value_modifier(
+        "test_step_size", modifier=lambda idx: pd.Series(pd.Timedelta(days=9), index=idx)
+    )
+    value = pipeline(index)
+    evens = value.iloc[lambda x: x.index % 2 == 0]
+    odds = value.iloc[lambda x: x.index % 2 == 1]
+    assert np.all(evens == pd.Timedelta(days=7))
+    assert np.all(odds == pd.Timedelta(days=5))
+
+    manager.register_value_modifier(
+        "test", modifier=lambda idx: pd.Series(pd.Timedelta(days=0.5), index=idx)
+    )
+    value = pipeline(index)
+    assert np.all(value == pd.Timedelta(days=2))
